very naive async error rethrow

+++ src/src/components/User/User.tsx
@@ -21,23 +21,49 @@ const User: React.FC<RouteComponentProps<UserRoute>> = ({
   history
 }) => {
   const [user, setUser] = useState<User | null>(null);
-  const [logs, setLogs] = useState<Log[]>([]);
-  const [userAccessList, setUserAccessList] = useState<LockAccess[]>([]);
+  const [logs, setLogs] = useState<APIResponse<Log> | null>(null);
+  const [locks, setLocks] = useState<APIResponse<Lock> | null>(null);
+  const [userAccessList, setUserAccessList] = useState<APIResponse<LockAccess> | null>(null);
+
+  const accessPagination = usePagination(userAccessList, setUserAccessList);
+  const logsPagination = usePagination(logs, setLogs);
+
+  const [loadError, setLoadError] = useState<Error | null>(null);
   const [showUserForm, setShowUserForm] = useState<boolean>(false);
   const toggleOverlay = useCallback(
     (type: 'open' | 'close') => () => setShowUserForm(type === 'open'),
     [setShowUserForm],
   );
   const deleteUser = useCallback(async () => {
-    const response = await API.delete(`users/${uId}/`);
-    if (response.ok) {
-      alert('Пользователь удален')
-      history.push('/users');
-    } else {
-      alert('Ошибка при удалении пользователя');
-      console.log(await response.text());
+    if (!confirm('Удалить пользователя?')) return;
+    try {
+      const response = await API.delete(`users/${uId}/`);
+      if (response.ok) {
+        alert('Пользователь удален');
+        history.push('/users');
+      } else {
+        alert('Ошибка при удалении пользователя');
+        console.log(await response.text());
+      }
+    } catch (error) {
+      setLoadError(error);
     }
   }, [user]);
+
+  const deleteAccess: deleteFn = async (id) => {
+    if (!confirm('Удалить доступ?')) return;
+    try {
+      const response = await API.delete(`accesses/${id}`);
+      if (!response.ok) {
+        alert('Ошибка при удалении доступа');
+      } else {
+        alert('Доступ удален');
+        getUserAccessList();
+      }
+    } catch (error) {
+      setLoadError(error);
+    }
+  }
 
   const userName = useMemo(() => {
     return user ? `#${user.u_id}: ${user.first_name} ${user.last_name}` : '';
@@ -45,17 +71,25 @@ const User: React.FC<RouteComponentProps<UserRoute>> = ({
 
   const saveUser = useCallback(
     async (data: FormData) => {
-      const response = await updateUser(uId, data);
-      alert('Пользователь обновлен');
-      setUser(response);
-      setShowUserForm(false);
+      try {
+        const response = await updateUser(uId, data);
+        alert('Пользователь обновлен');
+        setUser(response);
+        setShowUserForm(false);
+      } catch (error) {
+        setLoadError(error);
+      }
     },
     [uId],
   );
 
   async function getUserAccessList() {
-    const json = await apiGet<LockAccess>(`accesses/?u_id=${uId}`);
-    setUserAccessList(json);
+    try {
+      const json = await apiGet<LockAccess>(`accesses/?u_id=${uId}`);
+      setUserAccessList(json);
+    } catch (error) {
+      setLoadError(error);
+    }
   }
 
   const saveAccess = async (data: FormData) => {
@@ -65,8 +99,9 @@ const User: React.FC<RouteComponentProps<UserRoute>> = ({
       alert('Доступ добавлен');
       getUserAccessList();
       setShowAccessForm(false);
-    } catch (e) {
-      console.log(e);
+    } catch (error) {
+      console.log(error);
+      setLoadError(error);
     }
   }
 
@@ -86,10 +121,19 @@ const User: React.FC<RouteComponentProps<UserRoute>> = ({
       setLogs(json.results);
     }
 
-    getUser();
-    getUserAccessList();
-    getLogList();
+    try {
+      getUser();
+      getUserAccessList();
+      getLogList();
+      getLocks();
+    } catch (error) {
+      setLoadError(error);
+    }
   }, [uId]);
+
+  if (loadError) {
+    throw loadError;
+  }
 
   return (
     <div className="User Layout Layout_columns_2">
